First pass at modifying first round to collect sample counts when using client level dp with weighted fedavg

+++ fl4health/client_managers/base_sampling_manager.py
@@ -10,7 +10,7 @@ class BaseSamplingManager(SimpleClientManager):
 
     def sample(
         self,
-        sample_fraction: int,
+        sample_fraction: float,
         min_num_clients: Optional[int] = None,
         criterion: Optional[Criterion] = None,
     ) -> List[ClientProxy]:
